+++ src/plugins/profile/tests/status.js
@@ -34,7 +34,7 @@ describe('The Controlbox', function () {
                     <show>dnd</show>
                     <priority>0</priority>
                     <x xmlns="${Strophe.NS.VCARD_UPDATE}"></x>
-                    <c hash="sha-1" node="https://conversejs.org" ver="0cez5x5tbZP5NuLq2WClRoMmeKw=" xmlns="http://jabber.org/protocol/caps"/>
+                    <c hash="sha-1" node="https://conversejs.org" ver="LtdUA9LYgaH8dsa9k2gMJKk4fY0=" xmlns="http://jabber.org/protocol/caps"/>
                 </presence>`);
                 const view = await u.waitUntil(() => document.querySelector('converse-user-profile'));
                 const first_child = view.querySelector('.xmpp-status span:first-child');
@@ -65,7 +65,7 @@ describe('The Controlbox', function () {
                     <status>I am happy</status>
                     <priority>0</priority>
                     <x xmlns="${Strophe.NS.VCARD_UPDATE}"></x>
-                    <c hash="sha-1" node="https://conversejs.org" ver="0cez5x5tbZP5NuLq2WClRoMmeKw=" xmlns="http://jabber.org/protocol/caps"/>
+                    <c hash="sha-1" node="https://conversejs.org" ver="LtdUA9LYgaH8dsa9k2gMJKk4fY0=" xmlns="http://jabber.org/protocol/caps"/>
                 </presence>`);
 
                 const view = await u.waitUntil(() => document.querySelector('converse-user-profile'));